# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /buses:
    get:
      operationId: listBuses
      summary: List buses (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusesPage"
    post:
      operationId: createBus
      summary: Create a bus (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
  /buses/{id}:
    get:
      operationId: getBus
      summary: Get a bus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
    patch:
      operationId: updateBus
      summary: Update a bus (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bus"
    delete:
      operationId: deleteBus
      summary: Delete a bus (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DRIVERS ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── AUTH ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [credential, password]
      properties:
        credential:
          type: string
          description: Email or phone
        password:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── USER ────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, phone, role, walletBalance, isVerified, isBlocked, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, driver, admin]
        walletBalance:
          type: number
        isVerified:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Bus:
      type: object
      required: [id, plateNumber, capacity, model, isActive, createdAt]
      properties:
        id:
          type: integer
        plateNumber:
          type: string
        capacity:
          type: integer
        model:
          type: string
        currentLatitude:
          type: ["number", "null"]
        currentLongitude:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    BusInput:
      type: object
      required: [plateNumber, capacity, model]
      properties:
        plateNumber:
          type: string
        capacity:
          type: integer
        model:
          type: string
        isActive:
          type: boolean
          default: true
    BusUpdate:
      type: object
      properties:
        plateNumber:
          type: string
        capacity:
          type: integer
        model:
          type: string
        isActive:
          type: boolean
    BusesPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Bus"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── DRIVER ──────────────────────────────────────────────────────────────
...